pyproject.toml
[tool.poetry]
name = "there-gnome"
version = "0.1.0"
description = ""
authors = ["RJVB09 <rjvb09@ziggo.nl>"]
readme = "README.md"
packages = [{include = "there_gnome", from = "src"}]

[tool.poetry.dependencies]
python = "^3.13"
pyvisa = "^1.14.1"
pyserial = "^3.5"
pyvisa-py = "^0.7.2"
matplotlib = "^3.9.3"
pyqtgraph = "^0.13.7"
pyqt6 = "^6.7.1"
pygame = "^2.6.1"
sounddevice = "^0.5.1"
pysinewave = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
